pyproject.toml
[tool.poetry]
name = "sahayak"
version = "0.1.0"
description = "Multi-modal AI application with speech and face recognition"
authors = ["Your Name <your.email@example.com>"]
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
streamlit = "1.28.1"

# Exact versions from requirements.txt
opencv-python = "4.8.1.78"
face-recognition = "1.3.0"
numpy = "^1.26.4"
pandas = "2.0.3"
Pillow = "10.0.0"
dlib = "19.24.2"
cmake = "3.27.2"

# Additional dependencies from your previous setup
python-dotenv = "^1.0.0"
google-cloud-speech = "^2.21.0"
google-cloud-translate = "^3.21.1"
pyaudio = "^0.2.13"
google-auth = "^2.25.2"
google-api-core = "^2.15.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"